+++ p_4/e_2/components/distance_sensor/include/distance_sensor_reading.h
@@ -7,14 +7,14 @@ typedef struct distance_sensor_reading {
     uint64_t reading_id;
     int      adc_reading;
     int      voltage_mv;
-    uint8_t  distance_mm;
+    int      distance_mm;
 } distance_sensor_reading_t;
 
 void make_distance_sensor_reading(
     distance_sensor_reading_t *reading,
     int adc_reading,
     int voltage_mv,
-    uint8_t distance_mm
+    int distance_mm
 );
 
 #endif // __DISTANCE_SENSOR_READING_H__
